get db values in membersModal index instead of passing through as props
